+++ EDDI/ChangeLog.md
@@ -17,6 +17,7 @@ Full details of the variables available for each noted event, and VoiceAttack in
     * Fixed a bug with speech queues that could in some circumstances cause EDDI to crash in the middle of combat.
     * Added new event `Commander loading`, triggered at the very beginning of loading a game.
     * Added new event `Statistics`, triggered while loading a game.
+    * Added new event `Powerplay`, triggered while loading the game (if pledged).
     * Updated the `Commander continued` event with new properties `startlanded` and `startdead` (true if starting the game either landed or dead, respectively).
     * Updated the `System report` script to enhance the description of powerplay status.
     * Updated the `Engineer progressed` event to stay silent by default when written at startup (with empty values, signaling that engineer data has been loaded).